Update touch-panel-esp32.md
Previously there was a "Firmware Installation" main heading near the end of the page, with 18 steps including firmware flashing, configuring wifi (or not - if using ethernet), finding its IP address, getting device connected to MQTT etc. I have broken this up into more readable chunks / headings, and changed the main heading to "Setting up your Touch Panel". This remains near the bottom of the doc.

Also wrote a section on flashing the 86V / 86S.

I have then broken out "How does it work" - which was a subheading near the top of the doc, into its own main heading, and filled this section out. New main heading "How to communicate with your touch panel over MQTT", which includes an overview of what needs to be done to get it working. Then drills down into describing the /conf /stat /cmnd topics etc.

I feel the new "How to communicate..." section, which is still near the top, might be in the wrong place, and may be a little repetitive. I may try to re-factor the text layout to provide a more linear way to read through, but for now I think it's marginally better so hopefully the re-arranging of the deck-chairs isn't too pointless??

Cheers!